requirements.txt
beautifulsoup4==4.11.1
bs4==0.0.1
discord-ext-slash==0.9.2
discord.py==1.7.3
lxml==4.9.2
pip==23.2.1
requests==2.28.2
sympy==1.11.1